Reproducible interface is implemented in Animal; fixed NumberFormatException and rename getProperty method in GameFieldSettings; GameObject moved to service package; added toMate and probableNumberOfYoung methods to IslandRandom
